Se agrego AsNoTracking() en el endpoint patch

+++ MagicVilla_API/Modelos/Dto/VillaDto.cs
@@ -12,7 +12,7 @@ namespace MagicVilla_API.Modelos.Dto
         public string Detalle { get; set; }
         [Required]
         public double Tarifa { get; set; }  
-        public int MetrosCuadrados { get; set; }
+        public double MetrosCuadrados { get; set; }
         public int Ocupantes { get; set; }
         public string ImagenUrl { get; set; }
         public string Amenidad { get; set; }
